requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.2
Flask-HTTPAuth==4.8.0
fonttools==4.50.0
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.10.0
Werkzeug==3.0.1
yarl==1.9.4
